comboctl: Fix PairingSessionTest race conditions
Certain test states that affect the behavior of getNextSequenceEntry()
were subject to race conditions. Fix this by doing the following:

1. Prevent send() and receive() from running in parallel
2. Replace the race condition prone barrier channel with a condition
   variable (coroutine condition variables do not exist in kotlinx
   coroutines, so a custom implementation is used)

Also, expectedEndOfSequenceReached if block calls barrier.receive(),
which is also subject to race conditions - the channel might contain
a message when this if-block is entered. This would make the receive()
call incorrectly wake up. The intent is to suspend the coroutine
indefinitely until it is cancelled, so the right approach is to call
awaitCancellation() instead.